Ticket 219 - Conversion of integer variable to BigInteger

+++ base/common/src/com/netscape/cms/servlet/csadmin/CertUtil.java
@@ -357,15 +357,13 @@ public class CertUtil {
             if (type.equals("selfsign")) {
                 CMS.debug("Creating local certificate... issuerdn=" + dn);
                 CMS.debug("Creating local certificate... dn=" + dn);
-                info = CryptoUtil.createX509CertInfo(x509key, serialNo.intValue(), dn, dn, date,
-                        date, keyAlgorithm);
+                info = CryptoUtil.createX509CertInfo(x509key, serialNo, dn, dn, date, date, keyAlgorithm);
             } else {
                 String issuerdn = config.getString("preop.cert.signing.dn", "");
                 CMS.debug("Creating local certificate... issuerdn=" + issuerdn);
                 CMS.debug("Creating local certificate... dn=" + dn);
 
-                info = CryptoUtil.createX509CertInfo(x509key,
-                        serialNo.intValue(), issuerdn, dn, date, date, keyAlgorithm);
+                info = CryptoUtil.createX509CertInfo(x509key, serialNo, issuerdn, dn, date, date, keyAlgorithm);
             }
             CMS.debug("Cert Template: " + info.toString());
