+++ base/util/src/com/netscape/cmsutil/crypto/CryptoUtil.java
@@ -712,7 +712,7 @@ public class CryptoUtil {
     }
 
     public static X509CertInfo createX509CertInfo(X509Key x509key,
-            int serialno, String issuername, String subjname,
+            BigInteger serialno, String issuername, String subjname,
             Date notBefore, Date notAfter)
             throws IOException,
                 CertificateException,
@@ -728,12 +728,12 @@ public class CryptoUtil {
     }
 
     public static X509CertInfo createX509CertInfo(X509Key x509key,
-            int serialno, String issuername, String subjname,
+            BigInteger serialno, String issuername, String subjname,
             Date notBefore, Date notAfter, String alg)
             throws IOException,
-                CertificateException,
-                InvalidKeyException,
-                NoSuchAlgorithmException {
+            CertificateException,
+            InvalidKeyException,
+            NoSuchAlgorithmException {
         X509CertInfo info = new X509CertInfo();
 
         info.set(X509CertInfo.VERSION, new
